arw {mvoutlier}	R Documentation

Adaptive reweighted estimator for multivariate location and scatter

Description

Adaptive reweighted estimator for multivariate location and scatter with hard-rejection weights. The multivariate outliers are defined according to the supremum of the difference between the empirical distribution function of the robust Mahalanobis distance and the theoretical distribution function.

Usage

arw(x, m0, c0, alpha, pcrit)

Arguments

x 	 Dataset (n x p)

m0 	 Initial location estimator (1 x p)

c0 	 Initial scatter estimator (p x p)

alpha 	 Maximum thresholding proportion (optional scalar, default: alpha = 0.025)

pcrit 	 Critical value obtained by simulations (optional scalar, default value obtained from simulations)

Details

At the basis of initial estimators of location and scatter, the function arw performs a reweighting step to adjust the threshold for outlier rejection. The critical value pcrit was obtained by simulations using the MCD estimator as initial robust covariance estimator. If a different estimator is used, pcrit should be changed and computed by simulations for the specific dimensions of the data x.

Value

m 	 Adaptive location estimator (p x 1)

c 	 Adaptive scatter estimator (p x p)

cn 	 Adaptive threshold ("adjusted quantile")

w 	 Weight vector (n x 1)

Author(s)

Moritz Gschwandtner <e0125439@student.tuwien.ac.at>
Peter Filzmoser <P.Filzmoser@tuwien.ac.at> http://www.statistik.tuwien.ac.at/public/filz/

References

P. Filzmoser, R.G. Garrett, and C. Reimann. Multivariate outlier detection in exploration geochemistry. Computers & Geosciences, 31:579-587, 2005.

Examples

x <- cbind(rnorm(100), rnorm(100))
arw(x, apply(x,2,mean), cov(x))

[Package mvoutlier version 1.4 Index]