bssbot {mvoutlier}	R Documentation

Bottom Layer of the BSS Data

Description

The BSS data were collected in agrigultural soils from Northern Europe. from an area of about 1,800,000 km2. 769 samples on an iregular grid were taken in two different layers, the top layer (0-20cm) and the bottom layer. This dataset contains the bottom layer of the BSS data. It has 46 variables, including x and y coordinates.

Usage

data(bssbot)

Format

A data frame with 768 observations on the following 46 variables.

ID

a numeric vector

CNo

a numeric vector

XCOO

x coordinates: a numeric vector

YCOO

y coordinates: a numeric vector

SiO2_B

a numeric vector

TiO2_B

a numeric vector

Al2O3_B

a numeric vector

Fe2O3_B

a numeric vector

MnO_B

a numeric vector

MgO_B

a numeric vector

CaO_B

a numeric vector

Na2O_B

a numeric vector

K2O_B

a numeric vector

P2O5_B

a numeric vector

SO3_B

a numeric vector

Cl_B

a numeric vector

F_B

a numeric vector

LOI_B

a numeric vector

As_B

a numeric vector

Ba_B

a numeric vector

Bi_B

a numeric vector

Ce_B

a numeric vector

Co_B

a numeric vector

Cr_B

a numeric vector

Cs_B

a numeric vector

Cu_B

a numeric vector

Ga_B

a numeric vector

Hf_B

a numeric vector

La_B

a numeric vector

Mo_B

a numeric vector

Nb_B

a numeric vector

Ni_B

a numeric vector

Pb_B

a numeric vector

Rb_B

a numeric vector

Sb_B

a numeric vector

Sc_B

a numeric vector

Sn_B

a numeric vector

Sr_B

a numeric vector

Ta_B

a numeric vector

Th_B

a numeric vector

U_B

a numeric vector

V_B

a numeric vector

W_B

a numeric vector

Y_B

a numeric vector

Zn_B

a numeric vector

Zr_B

a numeric vector

Source

BSS Project in Northern Europe

References

Reimann C, Siewers U, Tarvainen T, Bityukova L, Eriksson J, Gilucis A, Gregorauskiene V, Lukashev VK, Matinian NN, Pasieczna A. Agricultural Soils in Northern Europe: A Geochemical Atlas. Geologisches Jahrbuch, Sonderhefte, Reihe D, Heft SD 5, Schweizerbart'sche Verlagsbuchhandlung, Stuttgart, 2003.

Examples

data(bssbot)
# classical versus robust correlation
cor.plot(log(bssbot[, "Al2O3_B"]), log(bssbot[, "Na2O_B"]))

[Package mvoutlier version 1.4 Index]