bsstop {mvoutlier}	R Documentation

Top Layer of the BSS Data

Description

The BSS data were collected in agrigultural soils from Northern Europe. from an area of about 1,800,000 km2. 769 samples on an iregular grid were taken in two different layers, the top layer (0-20cm) and the bottom layer. This dataset contains the top layer of the BSS data. It has 46 variables, including x and y coordinates.

Usage

data(bsstop)

Format

A data frame with 768 observations on the following 46 variables.

ID

a numeric vector

CNo

a numeric vector

XCOO

x coordinates: a numeric vector

YCOO

y coordinates: a numeric vector

SiO2_T

a numeric vector

TiO2_T

a numeric vector

Al2O3_T

a numeric vector

Fe2O3_T

a numeric vector

MnO_T

a numeric vector

MgO_T

a numeric vector

CaO_T

a numeric vector

Na2O_T

a numeric vector

K2O_T

a numeric vector

P2O5_T

a numeric vector

SO3_T

a numeric vector

Cl_T

a numeric vector

F_T

a numeric vector

LOI_T

a numeric vector

As_T

a numeric vector

Ba_T

a numeric vector

Bi_T

a numeric vector

Ce_T

a numeric vector

Co_T

a numeric vector

Cr_T

a numeric vector

Cs_T

a numeric vector

Cu_T

a numeric vector

Ga_T

a numeric vector

Hf_T

a numeric vector

La_T

a numeric vector

Mo_T

a numeric vector

Nb_T

a numeric vector

Ni_T

a numeric vector

Pb_T

a numeric vector

Rb_T

a numeric vector

Sb_T

a numeric vector

Sc_T

a numeric vector

Sn_T

a numeric vector

Sr_T

a numeric vector

Ta_T

a numeric vector

Th_T

a numeric vector

U_T

a numeric vector

V_T

a numeric vector

W_T

a numeric vector

Y_T

a numeric vector

Zn_T

a numeric vector

Zr_T

a numeric vector

Source

BSS Project in Northern Europe

References

Reimann C, Siewers U, Tarvainen T, Bityukova L, Eriksson J, Gilucis A, Gregorauskiene V, Lukashev VK, Matinian NN, Pasieczna A. Agricultural Soils in Northern Europe: A Geochemical Atlas. Geologisches Jahrbuch, Sonderhefte, Reihe D, Heft SD 5, Schweizerbart'sche Verlagsbuchhandlung, Stuttgart, 2003.

Examples

data(bsstop)
# classical versus robust correlation
cor.plot(log(bsstop[, "Al2O3_T"]), log(bsstop[, "Na2O_T"]))

[Package mvoutlier version 1.4 Index]