color.plot {mvoutlier}	R Documentation

Color Plot

Description

The function color.plot plots the (two-dimensional) data using different symbols according to the robust mahalanobis distance based on the mcd estimator with adjustment and using different colors according to the euclidean distances of the observations.

Usage

color.plot(x, quan=1/2, alpha=0.025, ...)

Arguments

x 	 two dimensional matrix or data.frame containing the data.

quan 	 amount of observations which are used for mcd estimations. has to be between 0.5 and 1, default ist 0.5

alpha 	 amount of observations used for calculating the adjusted quantile (see function arw).

... 	 additional graphical parameters

Details

The function color.plot plots the (two-dimensional) data using different symbols (see function symbol.plot) according to the robust mahalanobis distance based on the mcd estimator with adjustment and using different colors according to the euclidean distances of the observations. Blue is typical for a little distance, whereas red is the opposite. In addition four ellipsoids are drawn, on which mahalanobis distances are constant. These constant values correspond to the 25%, 50%, 75% and adjusted quantiles (see function arw) of the chi-square distribution (see Filzmoser et al., 2005).

Value

outliers 	 boolean vector of outliers

md 	 robust mahalanobis distances of the data

euclidean 	 euclidean distances of the observations according to the minimum of the data.

Author(s)

Moritz Gschwandtner <e0125439@student.tuwien.ac.at>
Peter Filzmoser <P.Filzmoser@tuwien.ac.at> http://www.statistik.tuwien.ac.at/public/filz/

References

P. Filzmoser, R.G. Garrett, and C. Reimann. Multivariate outlier detection in exploration geochemistry. Computers & Geosciences, 31:579-587, 2005.

See Also

symbol.plot, dd.plot, arw

Examples

# create data:
x <- cbind(rnorm(100), rnorm(100))
y <- cbind(rnorm(10, 5, 1), rnorm(10, 5, 1))
z <- rbind(x,y)
# execute:
color.plot(z, quan=0.75)

[Package mvoutlier version 1.4 Index]