dd.plot {mvoutlier}	R Documentation

Distance-Distance Plot

Description

The function dd.plot plots the classical mahalanobis distance of the data against the robust mahalanobis distance based on the mcd estimator. Different symbols (see function symbol.plot) and colours (see function color.plot) are used depending on the mahalanobis and euclidean distance of the observations (see Filzmoser et al., 2005).

Usage

dd.plot(x, quan=1/2, alpha=0.025, ...)

Arguments

x 	 matrix or data frame containing the data

quan 	 amount of observations which are used for mcd estimations. has to be between 0.5 and 1, default ist 0.5

alpha 	 amount of observations used for calculating the adjusted quantile (see function arw).

... 	 additional graphical parameters

Value

outliers 	 boolean vector of outliers

md.cla 	 mahalanobis distances of the observations based on classical estimators of location and scatter.

md.rob 	 mahalanobis distances of the observations based on robust estimators of location and scatter (mcd).

Author(s)

Moritz Gschwandtner <e0125439@student.tuwien.ac.at>
Peter Filzmoser <P.Filzmoser@tuwien.ac.at> http://www.statistik.tuwien.ac.at/public/filz/

References

P. Filzmoser, R.G. Garrett, and C. Reimann. Multivariate outlier detection in exploration geochemistry. Computers & Geosciences, 31:579-587, 2005.

See Also

symbol.plot, color.plot, arw, covPlot

Examples

# create data:
x <- cbind(rnorm(100), rnorm(100))
y <- cbind(rnorm(10, 3, 1), rnorm(10, 3, 1))
z <- rbind(x,y)
# execute:
dd.plot(z)
#
# Identify multivariate outliers for Co-Cu-Ni in humus layer of Kola data:
data(humus)
dd.plot(log(humus[,c("Co","Cu","Ni")]))

[Package mvoutlier version 1.4 Index]