moss {mvoutlier}	R Documentation

Moss Layer of the Kola Data

Description

The Kola Data were collected in the Kola Project (1993-1998, Geological Surveys of Finland (GTK) and Norway (NGU) and Central Kola Expedition (CKE), Russia). More than 600 samples in five different layers were analysed, this dataset contains the moss layer.

Usage

data(moss)

Format

A data frame with 598 observations on the following 34 variables.

ID

a numeric vector

XCOO

a numeric vector

YCOO

a numeric vector

Ag

a numeric vector

Al

a numeric vector

As

a numeric vector

B

a numeric vector

Ba

a numeric vector

Bi

a numeric vector

Ca

a numeric vector

Cd

a numeric vector

Co

a numeric vector

Cr

a numeric vector

Cu

a numeric vector

Fe

a numeric vector

Hg

a numeric vector

K

a numeric vector

Mg

a numeric vector

Mn

a numeric vector

Mo

a numeric vector

Na

a numeric vector

Ni

a numeric vector

P

a numeric vector

Pb

a numeric vector

Rb

a numeric vector

S

a numeric vector

Sb

a numeric vector

Si

a numeric vector

Sr

a numeric vector

Th

a numeric vector

Tl

a numeric vector

U

a numeric vector

V

a numeric vector

Zn

a numeric vector

Source

Kola Project (1993-1998)

References

Reimann C, Äyräs M, Chekushin V, Bogatyrev I, Boyd R, Caritat P de, Dutter R, Finne TE, Halleraker JH, Jæger Ø, Kashulina G, Lehto O, Niskavaara H, Pavlov V, Räisänen ML, Strand T, Volden T. Environmental Geochemical Atlas of the Central Barents Region. NGU-GTK-CKE Special Publication, Geological Survey of Norway, Trondheim, Norway, 1998.

Examples

data(moss)
# classical versus robust correlation:
cor.plot(log(moss[,"Al"]), log(moss[,"Na"]))

[Package mvoutlier version 1.4 Index]