pbb {mvoutlier}	R Documentation

BSS background Plot

Description

Plots the BSS background map

Usage

pbb(map = "bss.background", add.plot = FALSE, ...)

Arguments

map 	 List of coordinates. For the correct format see also help(kola.background)

add.plot 	 logical. If true background is added to an existing plot

... 	 additional plot parameters, see help(par)

Details

The list of coordinates is plotted as a polygon line.

Value

The plot is produced on the graphical device.

Author(s)

Peter Filzmoser <P.Filzmoser@tuwien.ac.at> http://www.statistik.tuwien.ac.at/public/filz/

References

Reimann C, Siewers U, Tarvainen T, Bityukova L, Eriksson J, Gilucis A, Gregorauskiene V, Lukashev VK, Matinian NN, Pasieczna A. Agricultural Soils in Northern Europe: A Geochemical Atlas. Geologisches Jahrbuch, Sonderhefte, Reihe D, Heft SD 5, Schweizerbart'sche Verlagsbuchhandlung, Stuttgart, 2003.

See Also

See also pkb

Examples

data(bss.background)
data(bsstop)
plot(bsstop$XCOO,bsstop$YCOO,col="red",pch=3)
pbb(add=TRUE)

[Package mvoutlier version 1.4 Index]