pkb {mvoutlier}	R Documentation

Kola background Plot

Description

Plots the Kola background map

Usage

pkb(map = "kola.background", which.map = c(1, 2, 3, 4), map.col = c(5, 1, 3, 4), map.lwd = c(2, 1, 2, 1), add.plot = FALSE, ...)

Arguments

map 	 List of coordinates. For the correct format see also help(kola.background)

which.map 	 which==1 ... plot project boundary # which==2 ... plot coast line # which==3 ... plot country borders # which==4 ... plot lakes and rivers

map.col 	 Map colors to be used

map.lwd 	 Defines linestyle of the background

add.plot 	 logical. if true background is added to an existing plot

... 	 additional plot parameters, see help(par)

Details

Is used by map.plot()

Author(s)

Peter Filzmoser <P.Filzmoser@tuwien.ac.at> http://www.statistik.tuwien.ac.at/public/filz/

References

Reimann C, Äyräs M, Chekushin V, Bogatyrev I, Boyd R, Caritat P de, Dutter R, Finne TE, Halleraker JH, Jæger Ø, Kashulina G, Lehto O, Niskavaara H, Pavlov V, Räisänen ML, Strand T, Volden T. Environmental Geochemical Atlas of the Central Barents Region. NGU-GTK-CKE Special Publication, Geological Survey of Norway, Trondheim, Norway, 1998.

Examples

example(map.plot)

[Package mvoutlier version 1.4 Index]