uni.plot {mvoutlier}	R Documentation

Univariate Presentation of Multivariate Outliers

Description

The function uni.plot plots each variable of x parallel in a one-dimensional scatter plot and in addition marks multivariate outliers.

Usage

uni.plot(x, symb=FALSE, quan=1/2, alpha=0.025, ...)

Arguments

x 	 matrix or data.frame containing the data.

symb 	 logical. if FALSE, only two colors and no special symbols are used. outliers are marked red. if TRUE different symbols (cross means big value, circle means little value) according to the robust mahalanobis distance based on the mcd estimator and different colors (red means big value, blue means little value) according to the euclidean distances of the observations are used.

quan 	 amount of observations which are used for mcd estimations. has to be between 0.5 and 1, default ist 0.5

alpha 	 amount of observations used for calculating the adjusted quantile (see function arw).

... 	 additional graphical parameters

Details

The function uni.plot shows the mutlivariate outliers in the single variables by one-dimensional scatter plots. If symb=FALSE (default), only two colors and no special symbols are used to mark multivariate outliers (the outliers are marked red). If symb=TRUE different symbols and colors are used. The symbols (cross means big value, circle means little value) are selected according to the robust mahalanobis distance based on the adjusted mcd estimator (see function symbol.plot) Different colors (red means big value, blue means little value) according to the euclidean distances of the observations (see function color.plot) are used. For details see Filzmoser et al. (2005).

Value

outliers 	 boolean vector of outliers

md 	 robust multivariate mahalanobis distances of the data

euclidean 	 (only if symb=TRUE) multivariate euclidean distances of the observations according to the minimum of the data.

Author(s)

Moritz Gschwandtner <e0125439@student.tuwien.ac.at>
Peter Filzmoser <P.Filzmoser@tuwien.ac.at> http://www.statistik.tuwien.ac.at/public/filz/

References

P. Filzmoser, R.G. Garrett, and C. Reimann. Multivariate outlier detection in exploration geochemistry. Computers & Geosciences, 31:579-587, 2005.

See Also

map.plot, symbol.plot, color.plot, arw

Examples

data(swiss)
uni.plot(swiss)
#
# Geostatistical data:
data(humus) # Load humus data
uni.plot(log(humus[, c("As", "Cd", "Co", "Cu", "Mg", "Pb", "Zn")]),symb=TRUE)

[Package mvoutlier version 1.4 Index]